requirements.txt
dj-database-url==0.5.0
Django==2.2.5
django-crispy-forms==1.7.2
gunicorn==19.9.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
whitenoise==3.2
